Cleaned random_color_jitter_generator code base.
